# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pub quizzes in Zagreb",
    "contact": {
      "name": "Dario",
      "email": "dm51987@fer.hr"
    },
    "license": {
      "name": "Creative Commons Zero v1.0 Universal"
    },
    "version": "1.2"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/quizzes": {
      "get": {
        "summary": "Get all the resources from collection",
        "description": "Returns a collection of all resources",
        "responses": {
          "200": {
            "description": "Found all quizzes successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          },
          "404": {
            "description": "Quiz not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Post a resource to collection",
        "description": "Posts a single resource to the collection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Quiz"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created quiz successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          },
          "404": {
            "description": "Quiz already found with that name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/quizzes/{id}": {
      "put": {
        "summary": "Put (update) resource from collection by id",
        "description": "Puts (updates) a single resource from the collection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Quiz"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated quiz successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          },
          "404": {
            "description": "Quiz not found by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete the resource from collection by id",
        "description": "Deletes a single collection resource",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted quiz by id successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          },
          "404": {
            "description": "Quiz not found by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get resource from collection by id",
        "description": "Returns a collection of resources that have the desired id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found quiz by id successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          },
          "404": {
            "description": "Quiz not found by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/quizzes/name/{name}": {
      "get": {
        "summary": "Get all the resources from collection by name",
        "description": "Returns a collection of all resources with desired name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found all quizzes successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          },
          "404": {
            "description": "Quiz not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/quizzes/team_number/{team_number}": {
      "get": {
        "summary": "Get all the resources from collection by team number",
        "description": "Returns a collection of all resources with desired team number",
        "parameters": [
          {
            "name": "team_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found all quizzes successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          },
          "404": {
            "description": "Quiz not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/quizzes/time/{time}": {
      "get": {
        "summary": "Get all the resources from collection by time",
        "description": "Returns a collection of all resources with desired time",
        "parameters": [
          {
            "name": "time",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found all quizzes successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          },
          "404": {
            "description": "Quiz not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Quiz": {
        "type": "object",
        "properties": {
          "@context": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LD-object"
            }
          },
          "id_kviz": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "popularity": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "team_number": {
            "type": "string"
          },
          "theme": {
            "type": "string"
          },
          "time": {
            "type": "string"
          },
          "rewards": {
            "type": "string"
          },
          "registrarion_fee": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "max_teams": {
            "type": "string"
          }
        }
      },
      "ResponseModel": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "response": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Quiz"
            }
          }
        }
      },
      "LD-object": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "http://schema.org/name"
          },
          "description": {
            "type": "string",
            "description": "http://schema.org/description"
          },
          "time": {
            "type": "string",
            "description": "http://schema.org/Time"
          },
          "contact": {
            "type": "string",
            "description": "http://schema.org/ContactPage"
          }
        }
      }
    }
  }
}